Brought in WW130 project work (I2C communication testing)
I had been working getting I2C slave comms (on the Seeed board) working with I2C master (on the WW130).

The main projects I worked on are i2c_slave_app and ww130_test.

I have also proposed a method of separating our WW projects from the Himax sceanrio_app projects.

+++ EPII_CM55M_APP_S/app/scenario_app/hello_world_freertos_tz_s_only/hardfault_handler.c
@@ -46,7 +46,7 @@ void HardFault_Handler(void) {
 		}
 		if (SAU->SFSR & SAU_SFSR_SFARVALID_Msk) {
 			/* SFARVALID: SFAR contain valid address that caused secure violation */
-			printf("Address that caused SAU violation is 0x%X.\r\n", SAU->SFAR);
+			printf("Address that caused SAU violation is 0x%X.\r\n", (int) SAU->SFAR);
 		}
 	}
 
@@ -65,7 +65,7 @@ void HardFault_Handler(void) {
 		if (SCB->CFSR & SCB_CFSR_BFARVALID_Msk) {
 			/* BFARVALID: BFAR contain valid address that caused secure violation */
 			printf("Address that caused secure bus violation is 0x%X.\r\n",
-					SCB->BFAR);
+					(int) SCB->BFAR);
 		}
 	}
 
@@ -94,9 +94,9 @@ void HardFault_Handler(void) {
     SCB->AIRCR =
         (SCB->AIRCR & ~SCB_AIRCR_VECTKEY_Msk) | (0x05FAUL << SCB_AIRCR_VECTKEY_Pos) | SCB_AIRCR_SYSRESETREQ_Msk;
 #else
-	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
-	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
-	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+	printf("SCB->CFSR:0x%08x\n", (int) SCB->CFSR);
+	printf("SCB->BFAR:0x%08x\n", (int) SCB->BFAR);
+	printf("SCB->HFSR:0x%08x\n", (int) SCB->HFSR);
 	for (;;) {
 	}
 #endif
@@ -115,9 +115,9 @@ void MemManage_Handler(void) {
 }
 void BusFault_Handler(void) {
 	printf("\r\nEntering BusFault_Handler interrupt!\r\n");
-	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
-	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
-	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+	printf("SCB->CFSR:0x%08x\n", (int) SCB->CFSR);
+	printf("SCB->BFAR:0x%08x\n", (int) SCB->BFAR);
+	printf("SCB->HFSR:0x%08x\n", (int) SCB->HFSR);
 	for (;;) {
 	}
 }
